+++ src/adplugins/adtextfile/datafile.cpp
@@ -135,7 +135,7 @@ datafile::getSpectrumCount( int /* fcn */ ) const
 }
 
 bool
-datafile::getSpectrum( int /* fcn */, int /* idx */, adcontrols::MassSpectrum&, uint32_t ) const
+datafile::getSpectrum( int /* fcn */, size_t /* idx */, adcontrols::MassSpectrum&, uint32_t ) const
 {
     return true;
 }
